refactor(ProfileMenu): break subcomponents into their own files
helps with separation of concerns and not clogging up lines in any single component file
