+++ src/components/Header/ProfileMenu/UserAvatar.tsx
@@ -1,0 +1,14 @@
+import { Avatar, type MantineSize } from "@mantine/core";
+
+interface UserAvatarProps {
+	username: string;
+	size?: MantineSize;
+}
+
+const UserAvatar = ({ username, size }: UserAvatarProps) => (
+	<Avatar variant="filled" color="cyan" radius="xl" size={size}>
+		{username[0].toUpperCase()}
+	</Avatar>
+);
+
+export default UserAvatar;
